ajout commentaires controller customer + first tentative delete reservation

+++ sql/tableInvoices.sql
@@ -3,11 +3,11 @@ CREATE TABLE `lemontagnard`.`invoices`
 (
     `idInvoice` INT AUTO_INCREMENT,
     `date` DATETIME NOT NULL,
-    `sumRooms` INT NOT NULL DEFAULT 0,
-    `sumExtras` INT NOT NULL DEFAULT 0,
-    `sumRestaurant` INT NOT NULL DEFAULT 0,
-    `advance` INT NOT NULL DEFAULT 0,
-    `discournt` INT NOT NULL DEFAULT 0,
+    `sumRooms` DOUBLE NOT NULL DEFAULT 0,
+    `sumExtras` DOUBLE NOT NULL DEFAULT 0,
+    `sumRestaurant` DOUBLE NOT NULL DEFAULT 0,
+    `advance` DOUBLE NOT NULL DEFAULT 0,
+    `discournt` DOUBLE NOT NULL DEFAULT 0,
     `idReservationI` INT NOT NULL,
     PRIMARY KEY (`idInvoice`),
     CONSTRAINT `FK_Invoice_Reservation` FOREIGN KEY (`idReservationI`) REFERENCES `reservationshotel` (`idReservation`)
